feat: split and improve contour min level selection in less time
